rename service
(cherry picked from commit 8661f10dc8af5bf3be7f0722a6185e086b9dc48f)
